+++ setup.cfg
@@ -30,4 +30,4 @@ test = pytest; pytest-cov; flake8
 max-complexity=10
 max-line-length=127
 ignore = W293
-exclude=test
+exclude=tests
